test(Validator): 객체 메서드 구현 순서와 테스트 케이스 순서 통일
